refactor: split ctf.services.participant into package (#889)
Resolve SonarCloud S104 by decomposing the monolithic participant service
into lifecycle, bulk_import, and queries submodules with a facade __init__
that preserves the public import path.
