+++ changelog.d/889.changed.md
@@ -1,0 +1,3 @@
+### Changed
+
+Internal refactor: split `ctf.services.participant` into a package (`lifecycle`, `bulk_import`, `queries`) to resolve SonarCloud file-length limit. Public API and behavior unchanged.
